Move paged-service to util
