+++ packages/sdk/src/index.ts
@@ -18,12 +18,12 @@ export * from './services/item-service';
 export * from './services/organization-members-service';
 export * from './services/organization-services-service';
 export * from './services/organizations-service';
-export * from './services/paged-service';
 export * from './services/secret-service';
 // export * from './services/share-service';
 export * from './services/templates-service';
 export * from './services/user-service';
 export * from './util/api-factory';
 export * from './util/find-connection-between';
+export * from './util/paged';
 import _cryppo from './services/cryppo-service';
 export const _cryppoService = _cryppo;
